requirements.txt
asgiref==3.8.1
Django==5.0.7
django-browser-reload==1.13.0
gunicorn==22.0.0
packaging==24.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
python-dotenv
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
pillow==10.3.0
